@@ -1,24 +1,23 @@
 apply plugin: 'com.github.ben-manes.versions'
 
 buildscript {
-    ext.kotlin_version = '1.3.41'
+    ext.kotlin_version = '1.9.22'
     repositories {
         google()
-        jcenter()
-        
+        mavenCentral()
+        gradlePluginPortal()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.5.0'
+        classpath 'com.android.tools.build:gradle:8.5.0'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
-        classpath 'com.github.ben-manes:gradle-versions-plugin:0.22.0'
+        classpath "com.github.ben-manes:gradle-versions-plugin:0.51.0"
     }
 }
 
 allprojects {
     repositories {
         google()
-        jcenter()
-        
+        mavenCentral()
     }
 }
 
@@ -27,13 +26,15 @@ task clean(type: Delete) {
 }
 
 dependencyUpdates.resolutionStrategy {
-    componentSelection { rules ->
-        rules.all { ComponentSelection selection ->
-            boolean rejected = ['alpha', 'beta', 'rc', 'cr', 'm', 'preview', 'b', 'ea'].any { qualifier ->
-                selection.candidate.version ==~ /(?i).*[.-]$qualifier[.\d-+]*/
+    componentSelection {
+        configureEach {
+            def isNonStable = { String version ->
+                ['jre', 'alpha', 'beta', 'rc', 'cr', 'm', 'preview', 'b', 'ea'].any { qualifier ->
+                    version ==~ /(?i).*[.-]$qualifier[.\d-+]*/
+                }
             }
-            if (rejected) {
-                selection.reject('Release candidate')
+            if (isNonStable(candidate.version) && !isNonStable(currentVersion)) {
+                reject('Release candidate')
             }
         }
     }
